perf(gpu): optimiza hash RPN y simplificador reduciendo sincronizaciones CPU
- Refactoriza cálculo de hash polinomial en TensorGeneticEngine usando operaciones tensoriales puras
- Cachea potencias del hash (_hash_powers) para evitar recálculo en bucles cerrados
- Elimina múltiples llamadas .item()/.any() que causaban sincronizaciones costosas CPU-GPU
- Pre-calcula IDs de operadores en GPUSymbolicSimplifier.__init__ en lugar de cada llamada a simplify_batch
- Elimina función helper _first_or reemplazándola por atributos de instancia

Estos cambios reducen drásticamente la sobrecarga de sincronización entre GPU y CPU en bucles de alto rendimiento.

+++ AlphaSymbolic/core/gpu/operators.py
@@ -418,7 +418,7 @@ class GPUOperators:
     def _validate_rpn_batch_custom(self, population: torch.Tensor, max_len: int) -> torch.Tensor:
         """Validate RPN batch for custom-length populations."""
         B, L = population.shape
-        arities = self.token_arity[population.clamp(0, self.token_arity.shape[0] - 1)]
+        arities = self.token_arity[population.clamp(0, self.token_arity.shape[0] - 1).long()]
         deltas = 1 - arities
         is_pad = (population == PAD_ID)
         deltas[is_pad] = 0
